requirements.txt
django
gunicorn
django-autoslug
